refactor: branching module organisation
