+++ pumpkin-solver/src/branching/value_selection/in_domain_interval.rs
@@ -1,6 +1,6 @@
 use super::InDomainSplit;
+use crate::branching::value_selection::ValueSelector;
 use crate::branching::SelectionContext;
-use crate::branching::ValueSelector;
 use crate::engine::predicates::predicate::Predicate;
 use crate::engine::variables::DomainId;
 use crate::predicate;
@@ -44,8 +44,8 @@ impl ValueSelector<DomainId> for InDomainInterval {
 mod tests {
     use super::InDomainInterval;
     use crate::basic_types::tests::TestRandom;
+    use crate::branching::value_selection::ValueSelector;
     use crate::branching::SelectionContext;
-    use crate::branching::ValueSelector;
     use crate::predicate;
 
     #[test]
