+++ pumpkin-solver/src/branching/value_selection/in_domain_split_random.rs
@@ -1,5 +1,5 @@
+use crate::branching::value_selection::ValueSelector;
 use crate::branching::SelectionContext;
-use crate::branching::ValueSelector;
 use crate::engine::predicates::predicate::Predicate;
 use crate::engine::variables::DomainId;
 use crate::predicate;
@@ -35,9 +35,9 @@ impl ValueSelector<DomainId> for InDomainSplitRandom {
 mod tests {
 
     use crate::basic_types::tests::TestRandom;
-    use crate::branching::InDomainSplitRandom;
+    use crate::branching::value_selection::InDomainSplitRandom;
+    use crate::branching::value_selection::ValueSelector;
     use crate::branching::SelectionContext;
-    use crate::branching::ValueSelector;
     use crate::predicate;
 
     #[test]
