+++ pumpkin-solver/src/branching/value_selection/out_domain_max.rs
@@ -1,5 +1,5 @@
+use crate::branching::value_selection::ValueSelector;
 use crate::branching::SelectionContext;
-use crate::branching::ValueSelector;
 use crate::engine::predicates::predicate::Predicate;
 use crate::engine::variables::DomainId;
 use crate::predicate;
@@ -21,9 +21,9 @@ impl ValueSelector<DomainId> for OutDomainMax {
 #[cfg(test)]
 mod tests {
     use crate::basic_types::tests::TestRandom;
-    use crate::branching::OutDomainMax;
+    use crate::branching::value_selection::OutDomainMax;
+    use crate::branching::value_selection::ValueSelector;
     use crate::branching::SelectionContext;
-    use crate::branching::ValueSelector;
     use crate::predicate;
 
     #[test]
