+++ pumpkin-solver/src/branching/value_selection/out_domain_median.rs
@@ -1,5 +1,5 @@
+use crate::branching::value_selection::ValueSelector;
 use crate::branching::SelectionContext;
-use crate::branching::ValueSelector;
 use crate::engine::predicates::predicate::Predicate;
 use crate::engine::variables::DomainId;
 use crate::predicate;
@@ -25,9 +25,9 @@ impl ValueSelector<DomainId> for OutDomainMedian {
 #[cfg(test)]
 mod tests {
     use crate::basic_types::tests::TestRandom;
-    use crate::branching::OutDomainMedian;
+    use crate::branching::value_selection::OutDomainMedian;
+    use crate::branching::value_selection::ValueSelector;
     use crate::branching::SelectionContext;
-    use crate::branching::ValueSelector;
     use crate::predicate;
 
     #[test]
