+++ pumpkin-solver/src/branching/value_selection/reverse_in_domain_split.rs
@@ -1,5 +1,5 @@
+use crate::branching::value_selection::ValueSelector;
 use crate::branching::SelectionContext;
-use crate::branching::ValueSelector;
 use crate::engine::predicates::predicate::Predicate;
 use crate::engine::variables::IntegerVariable;
 use crate::predicate;
@@ -36,11 +36,8 @@ impl<Var: IntegerVariable + Copy> ValueSelector<Var> for ReverseInDomainSplit {
 
 #[cfg(test)]
 mod tests {
+    use super::*;
     use crate::basic_types::tests::TestRandom;
-    use crate::branching::ReverseInDomainSplit;
-    use crate::branching::SelectionContext;
-    use crate::branching::ValueSelector;
-    use crate::predicate;
 
     #[test]
     fn test_returns_correct_literal() {
