+++ pumpkin-solver/src/branching/variable_selection/smallest.rs
@@ -1,10 +1,10 @@
 use log::warn;
 
 use super::VariableSelector;
-use crate::branching::Direction;
-use crate::branching::InOrderTieBreaker;
+use crate::branching::tie_breaking::Direction;
+use crate::branching::tie_breaking::InOrderTieBreaker;
+use crate::branching::tie_breaking::TieBreaker;
 use crate::branching::SelectionContext;
-use crate::branching::TieBreaker;
 use crate::engine::variables::DomainId;
 use crate::pumpkin_assert_eq_simple;
 
@@ -76,10 +76,8 @@ where
 
 #[cfg(test)]
 mod tests {
+    use super::*;
     use crate::basic_types::tests::TestRandom;
-    use crate::branching::SelectionContext;
-    use crate::branching::Smallest;
-    use crate::branching::VariableSelector;
 
     #[test]
     fn test_correctly_selected() {
